requirements.txt
absl-py==0.9.0
asgiref==3.2.3
astor==0.8.1
beautifulsoup4==4.9.1
boto3==1.14.7
botocore==1.17.7
cachetools==3.1.1
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
cycler==0.10.0
dicttoxml==1.7.4
Django==2.2.7
django-cors-headers==3.1.1
django-mptt==0.5.2
django-rest-framework==0.1.0
djangorestframework==3.10.3
djangorestframework-simplejwt==4.3.0
docutils==0.15.2
fuzzywuzzy==0.17.0
gast==0.3.3
ghostscript==0.6
google==2.0.2
google-api-core==1.21.0
google-api-python-client==1.7.11
google-auth==1.18.0
google-auth-httplib2==0.0.3
google-cloud==0.34.0
google-cloud-core==1.3.0
google-cloud-storage==1.29.0
google-cloud-tasks==1.5.0
google-cloud-vision==1.0.0
google-resumable-media==0.5.1
googleapis-common-protos==1.52.0
grpc-google-iam-v1==0.12.3
grpcio==1.29.0
gunicorn==20.0.4
h5py==2.10.0
html5lib==1.0.1
httplib2==0.18.1
idna==2.9
importlib-metadata==1.7.0
jmespath==0.10.0
json2xml==3.3.2
Keras==2.2.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Markdown==3.2.2
matplotlib==3.2.2
mock==4.0.2
numpy==1.16.2
opencv-python==4.1.2.30
pandas==0.25.3
Pillow==7.1.2
protobuf==3.12.2
psycopg2-binary==2.8.4
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycparser==2.19
pycryptodome==3.9.7
PyJWT==1.7.1
pyparsing==2.4.7
PyPDF2==1.26.0
python-dateutil==2.8.1
python-decouple==3.3
python-http-client==3.2.7
python-magic==0.4.18
pytz==2019.3
PyYAML==5.3.1
reportlab==3.5.42
requests==2.24.0
rsa==4.0
s3transfer==0.3.3
scipy==1.4.1
sendgrid==6.3.1
Shapely==1.6.4.post2
six==1.13.0
soupsieve==1.9.5
SQLAlchemy==1.3.10
sqlparse==0.3.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
uritemplate==3.0.0
urllib3==1.25.6
Wand==0.5.7
webencodings==0.5.1
Werkzeug==1.0.1
xhtml2pdf==0.2.4
xmltodict==0.11.0
zipp==3.1.0
